add script to generate plots for further analysis
Create a Python script to generate the required plots for the report.
The plots themselves will not be commited to this git repo since it is
not necessary. The plots can be generated on the local machine of anyone
who whishes to get them. Moreover, it makes little to no sense to track
temporary images using git.